requirement.txt.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

gTTS==2.5.3
pydub==0.25.1	

google-generativeai==0.8.0

aiofiles==23.2.1
httpx==0.27.2
pydantic==2.9.2